requirements.txt
google-generativeai>=0.5.0
youtube-transcript-api>=0.6.0
requests>=2.31.0
beautifulsoup4>=4.12.0 